Project permissions - refactoring, move to shared views

+++ UJCSystem/Vokabular.MainService.DataContracts/Clients/MainServiceProjectClient.cs
@@ -547,7 +547,7 @@ namespace Vokabular.MainService.DataContracts.Clients
             }
         }
 
-        public PagedResultList<UserGroupContract> GetUserGroupsByProject(int projectId, int start, int count, string query)
+        public PagedResultList<UserGroupContract> GetUserGroupsByProject(long projectId, int start, int count, string query)
         {
             try
             {
